Fix a bug in countChar
The pointer `s` is incremented in the while condition statement. Then,
the value of the char pointed by (alread-incremented) pointer `s` will be
checked inside the while block skipping the first character.

The test case is also updated to include the special case when the input
string starts with the delimeter character.